requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
requests==2.31.0
edge-tts==6.1.1
gtts==2.4.0
moviepy==1.0.3
pydub==0.25.1
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
yt-dlp==2023.10.13
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.101.0
google-generativeai==0.3.0
ollama==0.0.36
pydantic==2.4.2
schedule==1.2.0
openai==1.3.5
python-dateutil==2.8.2
librosa==0.10.0
diffusers==0.21.4
torch==2.0.1
transformers==4.33.2
pyttsx3==2.90
scipy==1.11.0
